fixed `init -r` default value. also bump deps.
When running `gut init` without giving it a root directory with
-r, gut would crash with a message saying something like
"invalid root folder 'RootDirectory::default()'". This is now fixed,
so that not giving -r to `gut init` will select the home directory
correctly.

This commit also bumps dependencies. Note that one dependency had
to be kept to not-the-newest version, as I kept running into
dependency issues.