Make --all-orgs flag global to work both before and after subcommands
Changed the --all-orgs flag to be a global argument using clap's 'global = true'
attribute. This allows the flag to be used in both positions:

  gut -A status           # Before subcommand (original)
  gut status -A           # After subcommand (new)
  gut status --all-orgs   # Long form after subcommand (new)

The flag will now appear in the help text for all subcommands and can be
placed anywhere on the command line while maintaining the same functionality.